+++ vripper-gui/src/main/kotlin/me/mnlr/vripper/gui/model/settings/ConnectionSettingsModel.kt
@@ -1,6 +1,7 @@
 package me.mnlr.vripper.gui.model.settings
 
 import javafx.beans.property.SimpleIntegerProperty
+import javafx.beans.property.SimpleLongProperty
 import tornadofx.*
 
 class ConnectionSettingsModel {
@@ -10,8 +11,8 @@ class ConnectionSettingsModel {
     val maxTotalThreadsProperty = SimpleIntegerProperty()
     var maxTotalThreads: Int by maxTotalThreadsProperty
 
-    val timeoutProperty = SimpleIntegerProperty()
-    var timeout: Int by timeoutProperty
+    val timeoutProperty = SimpleLongProperty()
+    var timeout: Long by timeoutProperty
 
     val maxAttemptsProperty = SimpleIntegerProperty()
     var maxAttempts: Int by maxAttemptsProperty
